+++ crates/tinyfs/src/path.rs
